feat(repo_identity): relocate pure helpers from nexus.registry (RDR-137 Phase 5.2b, nexus-tts0d.21)
Moves 6 pure helpers from nexus.registry into nexus.repo_identity:
_resolve_main_repo, _repo_identity (2-tuple), _repo_identity_with_main
(3-tuple), _safe_collection, _sanitise_owner_segment,
_resolve_repo_collection, list_sibling_collections.

nexus.registry re-exports every helper for one release-cycle of
import-path back-compat; new code should import from
nexus.repo_identity directly. RepoRegistry class stays in
nexus.registry pending nexus-tts0d.20.

Updated 15+ production call sites in src/nexus/ to import from
nexus.repo_identity. Updated 7 test files to retarget
monkeypatch.setattr('nexus.registry._repo_identity', ...) to
the new module so mocks intercept the right binding.

Helpers were never registry-coupled — they implement git-worktree-
stable repo identity (git rev-parse --git-common-dir) and the
RDR-103 conformant collection-naming rules. Both outlive the
registry deletion that nexus-tts0d.20 will land.

570 regression tests green locally.

Closes nexus-tts0d.21.
